+++ rocAL/include/pipeline/node.h
@@ -23,6 +23,8 @@ THE SOFTWARE.
 #pragma once
 #include <memory>
 #include <set>
+#include <array>
+#include <utility>
 
 #include "pipeline/graph.h"
 #include "meta_data/meta_data_graph.h"
@@ -49,6 +51,7 @@ class Node {
     const Roi2DCords *get_dst_roi() { return _outputs[0]->info().roi().get_2D_roi(); }
     void set_graph_id(int id) { _graph_id = id; }
     int get_graph_id() { return _graph_id; }
+    virtual std::string node_name() const { return ""; }
     const std::vector<Argument>& get_args_list() const { return _args; }
 
    protected:
@@ -64,4 +67,15 @@ class Node {
     std::vector<std::shared_ptr<Node>> _prev;   // Stores the reference to a list of previous Nodes
     int _graph_id = -1;
     std::vector<Argument> _args;
+    /**
+     * @brief Template function to set node arguments using variadic templates and fold expressions.
+     * 
+     * This function creates Argument objects for each argument-name pair and stores them in the node.
+     * It uses fold expressions to expand the parameter pack at compile time.
+     */
+    template <size_t N, size_t... Indices, typename... Args>
+    void set_node_arguments(const std::array<std::string, N>& arg_names, std::index_sequence<Indices ...>, Args... args) {
+        // Fold expression to create Argument object for each argument in the node
+        (this->_args.push_back(Argument(arg_names[Indices], std::forward<Args>(args))), ...);
+    }
 };